requirements.txt
Flask
flask_sqlalchemy
python-dotenv
Pillow
Pillow-Heif